+++ flutter_theoplayer_sdk/flutter_theoplayer_sdk/CHANGELOG.md
@@ -1,3 +1,13 @@
+[Unreleased]
+
+* Deprecated getX and setX methods on THEOplayer API in favor of properties.
+* Added new properties on THEOplayer API. (`muted`, `played`, `buffered`, `videoTracks`, etc... for the full list check the API.)
+* Changed `THEOplayer.allowBackgroundPlayback()` to `THEOplayer.allowBackgroundPlayback`. **(Breaking change)**.
+* Changed `THEOplayer.allowAutomaticPictureInPicture()` to `THEOplayer.allowAutomaticPictureInPicture`. **(Breaking change)**.
+* Changed `THEOplayer.isEnded()` to `THEOplayer.isEnded`. **(Breaking change)**.
+* Changed `THEOplayer.isPlaying()` to `THEOplayer.isPlaying`. **(Breaking change)**.
+* Changed `THEOplayer.isSeeking()` to `THEOplayer.isSeeking`. **(Breaking change)**.
+
 ## 8.2.0
 
 * Fixed an issue where play/pause buttons were not connected in the Picture-in-Picture window on Android.